add builds to client overview
