+++ ui/src/components/MobileClientBuild.css
@@ -1,0 +1,14 @@
+.build-view-list-group {
+  padding-left: 7px;
+  border-top: 0 none; }
+
+.build-view-status-icon {
+  padding-right: 12px; }
+
+.build-view-list-group .build-info {
+  margin-right: 15px; }
+
+.build-view-list-group .build-info.timestamp {
+  color: #757575;
+  padding-left: 5px;
+  font-size: 10px; }
