--- ui/src/styles/MobileClientBuild.css
@@ -1,34 +1,0 @@
-.build-view-header {
-    color: #757575;
-    font-size: 11px;
-    font-weight: 500;
-    padding-left: 20px;
-}
-
-.build-view-list-group {
-    padding-left: 7px;
-}
-
-.build-view-list-group-item {
-    font-size: 14px;
-    border-top: 0 none;
-}
-
-.build-view-list-group-item-description {
-    font-size: 12px;
-}
-
-.build-view-status-icon {
-    padding-right: 12px;
-}
-
-.build-view-build-info {
-    display: inline-block;
-}
-
-.build-view-timestamp {
-    color: #757575;
-    padding-left: 5px;
-    font-size: 10px;
-    display: inline-block;
-}
